Implemented simple persistent storage with SQLite.
Static variables' persistence is probably the most interesting
feature it offers. Note that variables are persisted on each call
to storage::persist-statics rather than on exist but it doesn't
seem to impact performance in a noticeable way (as benchmarked by
Siege). As a bonus this means less is lost if the web application
crashes.