http.tcl: (Breaking) Changed the handler API
Now route handlers are passed the channels to which to write their
resonses. (Earlier their return value was used as the response.)
This change breaks all code that uses http.tcl but should not
require much effort to fix. (See the changes to example.tcl.) It
enables jimhttp potentially to do things with keep-alive
connections as well as to integrate with other Tcl web servers
(for which passing around the channel is the common idiom).

Bumped http.tcl version to 0.12.0.