http.tcl: Option to not close channels
Now you can indicate that ::http::serve should not close the
channel (i.e., the socket) after a handler is done with it.
This is another step towards it being possible to use long-living
connections with jimhttp.

Restructured the dictionary ::http::routes to be ordered first by
the route then by the method rather than the other way around.

Bumped version to 0.14.0.