json.tcl: (Breaking) Change schema format
The default schema for every value in a JSON object is now given
as the value of the special key "*". It is overridden by the
schemas for particular keys. In the numberDictArrays mode
(by default), the schema for arrays is given via the special key
"N*". The presence of "*" instead of "N*" in a Tcl dict that would
otherwise have been serialized as an array forces its serialization
as an object.

Support for schemas with the format of "type1:type2:type3:..." has
been removed as limiting and a bad match for Tcl's syntax.

In ::json::parse the default value of the argument numberDictArrays
is now "1". While numberDictArrays 0 is often convenient when you
know the exact schema of the incoming data, the discrapency between
[parse] and [stringify] is a potential source of errors.

v2.0.0